+++ spezspellz/tests/__init__.py
@@ -1,0 +1,9 @@
+"""This file contains tests."""
+from .simple_test import SimpleTest
+from .upload_api_test import UploadAPITest
+
+
+__all__ = (
+    "SimpleTest",
+    "UploadAPITest"
+)
